Add files via upload
CantHide.sh is a versatile file analysis and extraction tool designed for uncovering hidden data within different file formats. The script intelligently identifies the type of the input file and applies the appropriate extraction method based on its format. It supports the extraction of Zip and 7-zip archives, as well as the retrieval of plain text data from binary files.

Purpose
The primary purpose of CantHide.sh is to simplify the process of extracting and analyzing data from various file types, making it an essential tool for users involved in data forensics, cybersecurity, or any field where uncovering hidden information is crucial.

Usage
To use the script, simply provide the target file as an argument:

bash
Copy code
./CantHide.sh target_file
The script will automatically detect the file type and perform the necessary extraction, revealing any hidden data within the file.